+++ app/ante/handler_options.go
@@ -17,7 +17,7 @@ import (
 	vestingtypes "github.com/tharsis/evmos/v4/x/vesting/types"
 )
 
-// HandlerOptions defines the list of module keepers required to run the Evmos
+// HandlerOptions defines the list of module keepers required to run the Torque
 // AnteHandler decorators.
 type HandlerOptions struct {
 	AccountKeeper   evmtypes.AccountKeeper