+++ ibc/utils.go
@@ -29,14 +29,14 @@ func GetTransferSenderRecipient(packet channeltypes.Packet) (
 
 	// validate the sender bech32 address from the counterparty chain
 	// and change the bech32 human readable prefix (HRP) of the sender to `evmos`
-	sender, err = evmos.GetEvmosAddressFromBech32(data.Sender)
+	sender, err = evmos.GetTorqueAddressFromBech32(data.Sender)
 	if err != nil {
 		return nil, nil, "", "", sdkerrors.Wrap(err, "invalid sender")
 	}
 
 	// validate the recipient bech32 address from the counterparty chain
 	// and change the bech32 human readable prefix (HRP) of the recipient to `evmos`
-	recipient, err = evmos.GetEvmosAddressFromBech32(data.Receiver)
+	recipient, err = evmos.GetTorqueAddressFromBech32(data.Receiver)
 	if err != nil {
 		return nil, nil, "", "", sdkerrors.Wrap(err, "invalid recipient")
 	}