+++ init.sh
@@ -22,7 +22,7 @@ evmosd config chain-id $CHAINID
 # if $KEY exists it should be deleted
 evmosd keys add $KEY --keyring-backend $KEYRING --algo $KEYALGO
 
-# Set moniker and chain-id for Evmos (Moniker can be anything, chain-id must be an integer)
+# Set moniker and chain-id for Torque (Moniker can be anything, chain-id must be an integer)
 evmosd init $MONIKER --chain-id $CHAINID
 
 # Change parameter token denominations to aevmos