+++ x/claims/genesis_test.go
@@ -26,7 +26,7 @@ type GenesisTestSuite struct {
 
 	ctx sdk.Context
 
-	app     *app.Evmos
+	app     *app.Torque
 	genesis types.GenesisState
 }