+++ x/recovery/keeper/keeper_test.go
@@ -33,7 +33,7 @@ type KeeperTestSuite struct {
 
 	ctx sdk.Context
 
-	app         *app.Evmos
+	app         *app.Torque
 	queryClient types.QueryClient
 }